requirements.txt
matplotlib==3.9.0
numpy==1.26.4
streamlit==1.35.0
tensorflow==2.15.0
opencv-python-headless
pillow==10.3.0
streamlit-option-menu==0.3.13